feat: add cli.ts file
Added a simple cli.ts file as a wrapper for the main entry, eggs.ts,
to help increase compatibility with various other tools. For example,
the 'land' CLI tool (which allows you to run command-line projects
without needing to install them first) requires that a project has a
file named 'cli.ts' to be able to run it. The overall naming convention
is for module entry points to be named 'mod.ts', command-line entries to
be named 'cli.ts', and serverless / deno deploy projects' entry points
to be named 'main.ts[x]'. Using 'eggs.ts' is... odd, but instead of
renaming the file I decided to wrap it to avoid breaking things. Thx